+++ bluetooth_low_energy_darwin/lib/bluetooth_low_energy_darwin.dart
@@ -1,9 +1,11 @@
 import 'package:bluetooth_low_energy_platform_interface/bluetooth_low_energy_platform_interface.dart';
 
-import 'src/my_bluetooth_low_energy.dart';
+import 'src/my_central_manager.dart';
+import 'src/my_peripheral_manager.dart';
 
 abstract class BluetoothLowEnergyDarwin {
   static void registerWith() {
-    BluetoothLowEnergy.instance = MyBluetoothLowEnergy();
+    CentralManager.instance = MyCentralManager();
+    PeripheralManager.instance = MyPeripheralManager();
   }
 }