+++ bluetooth_low_energy_darwin/lib/src/my_central_manager.dart
@@ -1,28 +1,47 @@
 import 'dart:async';
-import 'dart:typed_data';
 
 import 'package:bluetooth_low_energy_platform_interface/bluetooth_low_energy_platform_interface.dart';
+import 'package:flutter/foundation.dart';
 
 import 'my_api.dart';
-import 'my_bluetooth_low_energy_manager.dart';
 import 'my_gatt_characteristic2.dart';
 import 'my_gatt_descriptor2.dart';
 
-class MyCentralManager extends MyBluetoothLowEnergyManager
-    implements CentralManager, MyCentralManagerFlutterApi {
-  MyCentralManager()
-      : _api = MyCentralManagerHostApi(),
-        _discoveredController = StreamController.broadcast(),
-        _peripheralStateChangedController = StreamController.broadcast(),
-        _characteristicValueChangedController = StreamController.broadcast();
-
+class MyCentralManager extends CentralManager
+    implements MyCentralManagerFlutterApi {
   final MyCentralManagerHostApi _api;
+  BluetoothLowEnergyState _state;
+  final StreamController<BluetoothLowEnergyStateChangedEventArgs>
+      _stateChangedController;
   final StreamController<DiscoveredEventArgs> _discoveredController;
   final StreamController<PeripheralStateChangedEventArgs>
       _peripheralStateChangedController;
   final StreamController<GattCharacteristicValueChangedEventArgs>
       _characteristicValueChangedController;
 
+  MyCentralManager()
+      : _api = MyCentralManagerHostApi(),
+        _state = BluetoothLowEnergyState.unknown,
+        _stateChangedController = StreamController.broadcast(),
+        _discoveredController = StreamController.broadcast(),
+        _peripheralStateChangedController = StreamController.broadcast(),
+        _characteristicValueChangedController = StreamController.broadcast();
+
+  @override
+  BluetoothLowEnergyState get state => _state;
+  @protected
+  set state(BluetoothLowEnergyState value) {
+    if (_state == value) {
+      return;
+    }
+    _state = value;
+    final eventArgs = BluetoothLowEnergyStateChangedEventArgs(state);
+    _stateChangedController.add(eventArgs);
+  }
+
+  @override
+  Stream<BluetoothLowEnergyStateChangedEventArgs> get stateChanged =>
+      _stateChangedController.stream;
   @override
   Stream<DiscoveredEventArgs> get discovered => _discoveredController.stream;
   @override
@@ -33,9 +52,16 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
       get characteristicValueChanged =>
           _characteristicValueChangedController.stream;
 
+  Future<void> _throwWithoutState(BluetoothLowEnergyState state) async {
+    if (this.state != state) {
+      throw BluetoothLowEnergyError(
+        '$state is expected, but current state is ${this.state}.',
+      );
+    }
+  }
+
   @override
   Future<void> setUp() async {
-    await throwWithoutState(BluetoothLowEnergyState.unknown);
     final args = await _api.setUp();
     final stateArgs =
         MyBluetoothLowEnergyStateArgs.values[args.stateNumberArgs];
@@ -45,26 +71,26 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
 
   @override
   Future<void> startDiscovery() async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     await _api.startDiscovery();
   }
 
   @override
   Future<void> stopDiscovery() async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     await _api.stopDiscovery();
   }
 
   @override
   Future<void> connect(Peripheral peripheral) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     final peripheralHashCodeArgs = peripheral.hashCode;
     await _api.connect(peripheralHashCodeArgs);
   }
 
   @override
   Future<void> disconnect(Peripheral peripheral) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     final peripheralHashCodeArgs = peripheral.hashCode;
     await _api.disconnect(peripheralHashCodeArgs);
   }
@@ -74,7 +100,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
     Peripheral peripheral, {
     required GattCharacteristicWriteType type,
   }) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     final peripheralHashCodeArgs = peripheral.hashCode;
     final typeArgs = type.toArgs();
     final typeNumberArgs = typeArgs.index;
@@ -87,7 +113,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
 
   @override
   Future<int> readRSSI(Peripheral peripheral) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     final peripheralHashCodeArgs = peripheral.hashCode;
     final rssi = await _api.readRSSI(peripheralHashCodeArgs);
     return rssi;
@@ -95,7 +121,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
 
   @override
   Future<List<GattService>> discoverGATT(Peripheral peripheral) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     if (peripheral is! MyPeripheral) {
       throw TypeError();
     }
@@ -121,7 +147,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
   Future<Uint8List> readCharacteristic(
     GattCharacteristic characteristic,
   ) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     if (characteristic is! MyGattCharacteristic2) {
       throw TypeError();
     }
@@ -142,7 +168,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
     required Uint8List value,
     required GattCharacteristicWriteType type,
   }) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     if (characteristic is! MyGattCharacteristic2) {
       throw TypeError();
     }
@@ -166,7 +192,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
     GattCharacteristic characteristic, {
     required bool state,
   }) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     if (characteristic is! MyGattCharacteristic2) {
       throw TypeError();
     }
@@ -184,7 +210,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
 
   @override
   Future<Uint8List> readDescriptor(GattDescriptor descriptor) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     if (descriptor is! MyGattDescriptor2) {
       throw TypeError();
     }
@@ -205,7 +231,7 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
     GattDescriptor descriptor, {
     required Uint8List value,
   }) async {
-    await throwWithoutState(BluetoothLowEnergyState.poweredOn);
+    await _throwWithoutState(BluetoothLowEnergyState.poweredOn);
     if (descriptor is! MyGattDescriptor2) {
       throw TypeError();
     }
@@ -232,15 +258,15 @@ class MyCentralManager extends MyBluetoothLowEnergyManager
   void onDiscovered(
     MyPeripheralArgs peripheralArgs,
     int rssiArgs,
-    MyAdvertiseDataArgs advertiseDataArgs,
+    MyAdvertisementArgs advertisementArgs,
   ) {
     final peripheral = peripheralArgs.toPeripheral();
     final rssi = rssiArgs;
-    final advertiseData = advertiseDataArgs.toAdvertiseData();
+    final advertisement = advertisementArgs.toAdvertisement();
     final eventArgs = DiscoveredEventArgs(
       peripheral,
       rssi,
-      advertiseData,
+      advertisement,
     );
     _discoveredController.add(eventArgs);
   }